requirements.txt
# â”€â”€ Core dependencies (actively used) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
apscheduler>=3.10.1
python-telegram-bot>=20.0
litellm>=1.30.0
PyYAML>=6.0.2
beautifulsoup4>=4.12.3
playwright>=1.52.0

# â”€â”€ Planned / future features (not yet used in code) â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
# pydantic>=2.0.0          # Data validation (planned)
# fastapi>=0.100.0         # Web API interface (planned)
# uvicorn>=0.23.0          # ASGI server for FastAPI (planned)
# customtkinter>=5.2.0     # Desktop GUI (planned)
# pystray>=0.19.5          # System tray icon (planned)
# Pillow>=10.0.0           # Image processing for tray icon (planned)

